--- a/envs/monkey_zoo/blackbox/test_blackbox.py
+++ b/envs/monkey_zoo/blackbox/test_blackbox.py
@@ -10,7 +10,7 @@ from envs.monkey_zoo.blackbox.analyzers.zerologon_analyzer import ZerologonAnaly
 from envs.monkey_zoo.blackbox.config_templates.config_template import ConfigTemplate
 from envs.monkey_zoo.blackbox.config_templates.grouped.depth_1_a import Depth1A
 from envs.monkey_zoo.blackbox.config_templates.grouped.depth_1_b import Depth1B
-from envs.monkey_zoo.blackbox.config_templates.grouped.depth_4_a import Depth4A
+from envs.monkey_zoo.blackbox.config_templates.grouped.depth_3_a import Depth3A
 from envs.monkey_zoo.blackbox.gcp_test_machine_list import GCP_TEST_MACHINE_LIST
 from envs.monkey_zoo.blackbox.island_client.island_config_parser import IslandConfigParser
 from envs.monkey_zoo.blackbox.island_client.monkey_island_client import MonkeyIslandClient
@@ -155,5 +155,5 @@ class TestMonkeyBlackbox:
             log_handler=log_handler,
         ).run()
 
-    def test_depth_4_a(self, island_client):
-        TestMonkeyBlackbox.run_exploitation_test(island_client, Depth4A, "Depth4A test suite")
+    def test_depth_3_a(self, island_client):
+        TestMonkeyBlackbox.run_exploitation_test(island_client, Depth3A, "Depth4A test suite")