--- a/monkey/monkey_island/Pipfile
+++ b/monkey/monkey_island/Pipfile
@@ -38,6 +38,7 @@ pytest-cov = "*"
 isort = "==5.10.1"
 coverage = "*"
 vulture = "==2.3"
+tqdm = "*"  # Used in BB tests
 
 [requires]
 python_version = "3.7"