… envs/monkey_zoo/blackbox/config_templates/single_tests/hadoop.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/log4j_logstash.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/log4j_solr.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/log4j_tomcat.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/mssql.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/performance.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/powershell.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/powershell_credentials_reuse.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/smb_mimikatz.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/smb_pth.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/ssh.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/tunneling.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/wmi_mimikatz.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/wmi_pth.py
… envs/monkey_zoo/blackbox/config_templates/single_tests/zerologon.py
… envs/monkey_zoo/blackbox/test_blackbox_in_depth.py
… envs/monkey_zoo/blackbox/utils/config_generation_script.py
… envs/monkey_zoo/docs/fullDocs.md
… monkey/monkey_island/Pipfile.lock