# What does this PR do?

Fixes #1813

Add any further explanations here.
# Before and after:

Last line `Took X min` represents the time for the whole jenkins job. This gives a perspective of how long the setup process takes in comparison.

## Linux:

Unoptimized, Hadoop 3 failed:
```
FAILED ../envs/monkey_zoo/blackbox/test_blackbox.py::TestMonkeyBlackbox::test_hadoop_exploiter
===== 1 failed, 12 passed, 9 skipped, 1901 warnings in 1137.70s (0:18:57) ======
Black box tests failed!
Took 24 min
```
Unoptimized again, Hadoop 3 failed:
```
=========================== short test summary info ============================
FAILED ../envs/monkey_zoo/blackbox/test_blackbox.py::TestMonkeyBlackbox::test_hadoop_exploiter
===== 1 failed, 12 passed, 9 skipped, 1871 warnings in 1200.87s (0:20:00) ======
Black box tests failed!
Took 26 min
```
Optimized, a lot of failures:
```
-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=========================== short test summary info ============================
FAILED ../envs/monkey_zoo/blackbox/test_blackbox.py::TestMonkeyBlackbox::test_depth_1_a
FAILED ../envs/monkey_zoo/blackbox/test_blackbox.py::TestMonkeyBlackbox::test_depth_1_b
============ 2 failed, 1 passed, 991 warnings in 680.96s (0:11:20) =============
Black box tests failed!
Took 16 min
```

Windows:

Unoptimized, ssh and tunneling fails:
```
-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=========================== short test summary info ===========================
FAILED ..\..\infection_monkey\envs\monkey_zoo\blackbox\test_blackbox.py::TestMonkeyBlackbox::test_ssh_exploiter
FAILED ..\..\infection_monkey\envs\monkey_zoo\blackbox\test_blackbox.py::TestMonkeyBlackbox::test_tunneling
===== 2 failed, 11 passed, 9 skipped, 730 warnings in 1219.55s (0:20:19) ======
Took 34 min
```

Optimized, more tests fail:
```
=========================== short test summary info ===========================
FAILED ..\..\infection_monkey\envs\monkey_zoo\blackbox\test_blackbox.py::TestMonkeyBlackbox::test_depth_1_a
====== 1 failed, 5 passed, 1 skipped, 425 warnings in 737.06s (0:12:17) =======
Took [26 min]
```


## Testing Checklist

* [ ] Have you successfully tested your changes locally? Elaborate:
    > Tested by running the BB tests manually and from Jenkins